requirements.txt
requests>=2.28.0
cryptography>=41.0.0
numpy>=1.24.0
opencv-python>=4.8.0
urllib3>=2.0.0
python-dotenv>=1.0.0
